requirements.txt
pydantic>=2.0.0
pydantic-ai
python-dateutil>=2.8.2
httpx>=0.27.0
openai>=1.0.0
supabase>=2.10.0
python-dotenv>=1.0.1
google-genai>=0.2.0
pydantic-ai-slim[web]
uvicorn
mlflow
pandas